- resolves #140 - added a "level" option to only fetch a messages at or above a certain log level
