+++ src/ops/LogOps.js
@@ -1,0 +1,245 @@
+import { saveConnection } from '../api/AuthApi.js';
+import { printMessage } from './utils/Console.js';
+import { getCurrentTimestamp } from './utils/ExportImportUtils.js';
+import {
+  createAPIKeyAndSecret,
+  getAPIKeys,
+  getSources,
+} from '../api/LogApi.js';
+
+import storage from '../storage/SessionStorage.js';
+
+import * as LogApi from '../api/LogApi.js';
+
+const miscNoise = [
+  'text/plain',
+  'com.iplanet.dpro.session.operations.ServerSessionOperationStrategy',
+  'com.iplanet.dpro.session.SessionIDFactory',
+  'com.iplanet.dpro.session.share.SessionEncodeURL',
+  'com.iplanet.services.naming.WebtopNaming',
+  'com.iplanet.sso.providers.dpro.SSOProviderImpl',
+  'com.sun.identity.authentication.AuthContext',
+  'com.sun.identity.authentication.client.AuthClientUtils',
+  'com.sun.identity.authentication.config.AMAuthConfigType',
+  'com.sun.identity.authentication.config.AMAuthenticationManager',
+  'com.sun.identity.authentication.config.AMAuthLevelManager',
+  'com.sun.identity.authentication.config.AMConfiguration',
+  'com.sun.identity.authentication.jaas.LoginContext',
+  'com.sun.identity.authentication.modules.application.Application',
+  'com.sun.identity.authentication.server.AuthContextLocal',
+  'com.sun.identity.authentication.service.AMLoginContext',
+  'com.sun.identity.authentication.service.AuthContextLookup',
+  'com.sun.identity.authentication.service.AuthD',
+  'com.sun.identity.authentication.service.AuthUtils',
+  'com.sun.identity.authentication.service.DSAMECallbackHandler',
+  'com.sun.identity.authentication.service.LoginState',
+  'com.sun.identity.authentication.spi.AMLoginModule',
+  'com.sun.identity.delegation.DelegationEvaluatorImpl',
+  'com.sun.identity.idm.plugins.internal.AgentsRepo',
+  'com.sun.identity.idm.server.IdCachedServicesImpl',
+  'com.sun.identity.idm.server.IdRepoPluginsCache',
+  'com.sun.identity.idm.server.IdServicesImpl',
+  'com.sun.identity.log.spi.ISDebug',
+  'com.sun.identity.shared.encode.CookieUtils',
+  'com.sun.identity.sm.ldap.SMSLdapObject',
+  'com.sun.identity.sm.CachedSMSEntry',
+  'com.sun.identity.sm.CachedSubEntries',
+  'com.sun.identity.sm.DNMapper',
+  'com.sun.identity.sm.ServiceConfigImpl',
+  'com.sun.identity.sm.ServiceConfigManagerImpl',
+  'com.sun.identity.sm.SMSEntry',
+  'com.sun.identity.sm.SMSUtils',
+  'com.sun.identity.sm.SmsWrapperObject',
+  'oauth2',
+  'org.apache.http.client.protocol.RequestAuthCache',
+  'org.apache.http.impl.conn.PoolingHttpClientConnectionManager',
+  'org.apache.http.impl.nio.client.InternalHttpAsyncClient',
+  'org.apache.http.impl.nio.client.InternalIODispatch',
+  'org.apache.http.impl.nio.client.MainClientExec',
+  'org.apache.http.impl.nio.conn.ManagedNHttpClientConnectionImpl',
+  'org.apache.http.impl.nio.conn.PoolingNHttpClientConnectionManager',
+  'org.forgerock.audit.AuditServiceImpl',
+  'org.forgerock.oauth2.core.RealmOAuth2ProviderSettings',
+  'org.forgerock.openam.authentication.service.JAASModuleDetector',
+  'org.forgerock.openam.authentication.service.LoginContextFactory',
+  'org.forgerock.openam.blacklist.BloomFilterBlacklist',
+  'org.forgerock.openam.blacklist.CTSBlacklist',
+  'org.forgerock.openam.core.realms.impl.CachingRealmLookup',
+  'org.forgerock.openam.core.rest.authn.RestAuthCallbackHandlerManager',
+  'org.forgerock.openam.core.rest.authn.trees.AuthTrees',
+  'org.forgerock.openam.cors.CorsFilter',
+  'org.forgerock.openam.cts.CTSPersistentStoreImpl',
+  'org.forgerock.openam.cts.impl.CoreTokenAdapter',
+  'org.forgerock.openam.cts.impl.queue.AsyncResultHandler',
+  'org.forgerock.openam.cts.reaper.ReaperDeleteOnQueryResultHandler',
+  'org.forgerock.openam.headers.DisableSameSiteCookiesFilter',
+  'org.forgerock.openam.idrepo.ldap.DJLDAPv3Repo',
+  'org.forgerock.openam.rest.CsrfFilter',
+  'org.forgerock.openam.rest.restAuthenticationFilter',
+  'org.forgerock.openam.rest.fluent.CrestLoggingFilter',
+  'org.forgerock.openam.session.cts.CtsOperations',
+  'org.forgerock.openam.session.stateless.StatelessSessionManager',
+  'org.forgerock.openam.sm.datalayer.impl.ldap.ExternalLdapConfig',
+  'org.forgerock.openam.sm.datalayer.impl.ldap.LdapQueryBuilder',
+  'org.forgerock.openam.sm.datalayer.impl.SeriesTaskExecutor',
+  'org.forgerock.openam.sm.datalayer.impl.SeriesTaskExecutorThread',
+  'org.forgerock.openam.sm.datalayer.providers.LdapConnectionFactoryProvider',
+  'org.forgerock.openam.sm.file.ConfigFileSystemHandler',
+  'org.forgerock.openam.social.idp.SocialIdentityProviders',
+  'org.forgerock.openam.utils.ClientUtils',
+  'org.forgerock.opendj.ldap.CachedConnectionPool',
+  'org.forgerock.opendj.ldap.LoadBalancer',
+  'org.forgerock.secrets.keystore.KeyStoreSecretStore',
+  'org.forgerock.secrets.propertyresolver.PropertyResolverSecretStore',
+  'org.forgerock.secrets.SecretsProvider',
+];
+
+const journeysNoise = [
+  'org.forgerock.openam.auth.trees.engine.AuthTreeExecutor',
+];
+
+// eslint-disable-next-line no-unused-vars
+const journeys = [
+  'org.forgerock.openam.auth.nodes.SelectIdPNode',
+  'org.forgerock.openam.auth.nodes.ValidatedPasswordNode',
+  'org.forgerock.openam.auth.nodes.ValidatedUsernameNode',
+  'org.forgerock.openam.auth.trees.engine.AuthTreeExecutor',
+];
+
+const samlNoise = [
+  'com.sun.identity.cot.COTCache',
+  'com.sun.identity.plugin.configuration.impl.ConfigurationInstanceImpl',
+  'com.sun.identity.saml2.meta.SAML2MetaCache',
+  'com.sun.identity.saml2.profile.CacheCleanUpRunnable',
+  'org.apache.xml.security.keys.KeyInfo',
+  'org.apache.xml.security.signature.XMLSignature',
+  'org.apache.xml.security.utils.SignerOutputStream',
+  'org.apache.xml.security.utils.resolver.ResourceResolver',
+  'org.apache.xml.security.utils.resolver.implementations.ResolverFragment',
+  'org.apache.xml.security.algorithms.JCEMapper',
+  'org.apache.xml.security.algorithms.implementations.SignatureBaseRSA',
+  'org.apache.xml.security.algorithms.SignatureAlgorithm',
+  'org.apache.xml.security.utils.ElementProxy',
+  'org.apache.xml.security.transforms.Transforms',
+  'org.apache.xml.security.utils.DigesterOutputStream',
+  'org.apache.xml.security.signature.Reference',
+  'org.apache.xml.security.signature.Manifest',
+];
+
+// eslint-disable-next-line no-unused-vars
+const saml = [
+  'jsp.saml2.spAssertionConsumer',
+  'com.sun.identity.saml.common.SAMLUtils',
+  'com.sun.identity.saml2.common.SAML2Utils',
+  'com.sun.identity.saml2.meta.SAML2MetaManager',
+  'com.sun.identity.saml2.xmlsig.FMSigProvider',
+];
+
+const noise = miscNoise.concat(samlNoise).concat(journeysNoise);
+
+export function resolveLevel(levelName) {
+  const levels = ['FATAL', 'ERROR', 'WARN', 'INFO', 'DEBUG', 'TRACE', 'ALL'];
+  levels.splice(levels.indexOf(levelName) + 1, levels.length);
+  return levels;
+}
+
+export async function getLogSources() {
+  const sources = [];
+  await getSources()
+    .then((response) => {
+      response.data.result.forEach((item) => {
+        sources.push(item);
+      });
+    })
+    .catch((error) => {
+      printMessage(
+        `getSources ERROR: get log sources call returned ${error}}`,
+        'error'
+      );
+      return [];
+    });
+  return sources;
+}
+
+export async function tailLogs(source, levels, cookie) {
+  try {
+    const response = await LogApi.tail(source, cookie);
+    if (response.status < 200 || response.status > 399) {
+      printMessage(
+        `tail ERROR: tail call returned ${response.status}`,
+        'error'
+      );
+      return null;
+    }
+    // if (!cookie) {
+    //   await saveConnection();
+    // }
+    const logsObject = response.data;
+    let filteredLogs = [];
+    if (Array.isArray(logsObject.result)) {
+      filteredLogs = logsObject.result.filter(
+        (el) =>
+          !noise.includes(el.payload.logger) &&
+          !noise.includes(el.type) &&
+          levels.includes(el.payload.level)
+      );
+    }
+
+    filteredLogs.forEach((e) => {
+      printMessage(JSON.stringify(e.payload), 'data');
+    });
+
+    setTimeout(() => {
+      tailLogs(source, levels, logsObject.result.pagedResultsCookie);
+    }, 5000);
+    return null;
+  } catch (e) {
+    printMessage(`tail ERROR: tail data error - ${e}`, 'error');
+    return `tail ERROR: tail data error - ${e}`;
+  }
+}
+
+export async function provisionCreds() {
+  try {
+    let keyName = `frodo-${storage.session.getUsername()}`;
+    return getAPIKeys()
+      .then((response) => {
+        response.data.result.forEach((k) => {
+          if (k.name === keyName) {
+            // append current timestamp to name if the named key already exists
+            keyName = `${keyName}-${getCurrentTimestamp()}`;
+          }
+        });
+        return createAPIKeyAndSecret(keyName)
+          .then((resp) => {
+            if (resp.data.name !== keyName) {
+              printMessage(
+                `create keys ERROR: could not create log API key ${keyName}`,
+                'error'
+              );
+              return null;
+            }
+            printMessage(
+              `Created a new log API key [${keyName}] in ${storage.session.getTenant()}`
+            );
+            return resp.data;
+          })
+          .catch((error) => {
+            printMessage(
+              `create keys ERROR: create keys call returned ${error}`,
+              'error'
+            );
+            return null;
+          });
+      })
+      .catch((error) => {
+        printMessage(
+          `get keys ERROR: get keys call returned ${error}`,
+          'error'
+        );
+      });
+  } catch (e) {
+    printMessage(`create keys ERROR: create keys data error - ${e}`, 'error');
+    return null;
+  }
+}
